refactor(worker): move WorkerTestAccess out of the production header (5D review)
Review finding 2: WorkerTestAccess was fully defined in the production header, so
any translation unit could install the fault hook or read private worker state —
it replaced four test methods with a larger globally available backdoor rather
than removing the test surface.

Forward-declare it as a friend in the worker header and DEFINE it (with its
nested ProgressForTest) in a new test-only header, tests/workertestaccess.h,
included by the worker suite (and, later, the phase-5 integration runner).
Production code can no longer reach the worker's private state through it, and
the worker's public API keeps no *ForTest / Set* methods and no test data structs.

Review finding 3: fix three stale comments that 5C/5D falsified —
  * the W-THROW handler pin still said the catch-all decides ownership by
    "generation/state" and that AbortUpdate() forces counter completion; the
    catch is now an unconditional idempotent AbortUpdate() and failure is
    counter-independent;
  * the W-ER1 list-lane pin still described counter snapping in the list-failure
    branch (removed in 5C — it is the direct terminal path now);
  * FakePoeApiClient::Call::token still said real tokens are not asserted yet
    (they are, since 5B: W-TOKEN/W-IDENTITY).

Full ctest suite green (22/22).
