+++ src/itemsmanagerworker.h
@@ -100,20 +100,11 @@ public:
 
     void UpdateRequest(TabSelection type, const std::vector<ItemLocation> &locations);
 
-    // A snapshot of the completion counters (verification §2/§3), read on each
-    // StatusUpdate to pin that reported progress is monotonic (P-STATUS). Pure
-    // data; the sole test seam that reads it is WorkerTestAccess below.
-    struct ProgressForTest
-    {
-        size_t stashes_received;
-        size_t stashes_needed;
-        size_t characters_received;
-        size_t characters_needed;
-    };
-
     // The single test-only seam into the worker (network-redesign phase 5,
-    // verification §2). It reaches private state directly as a friend, so the
-    // worker's public API carries no standing *ForTest / Set* methods.
+    // verification §2). Only forward-declared here and DEFINED in a test-only
+    // header (tests/workertestaccess.h), so it grants the worker suite private
+    // access without adding any test surface — production code cannot reach it,
+    // and the worker's public API carries no standing *ForTest / Set* methods.
     friend class WorkerTestAccess;
 
     // The setting values are parameters because this runs on the parser
@@ -345,49 +336,4 @@ private:
     // per-fetch handle until the deferred sweep reclaims the completed ones.
     QCoro::Task<> m_update_task;
     std::vector<QCoro::Task<>> m_fetch_tasks;
-};
-
-// The one test-only seam into ItemsManagerWorker (network-redesign phase 5,
-// verification §2): a friend accessor that bundles every observation and
-// injection the worker suite needs — the injected read-only sweep observer and
-// one-shot fault hook, and read-only snapshots of the outstanding task handles
-// and the progress counters — instead of standing *ForTest / Set* methods on the
-// production API. Constructed only by tests; it reaches private state as a friend
-// and adds nothing to the worker's own interface.
-class WorkerTestAccess
-{
-public:
-    explicit WorkerTestAccess(ItemsManagerWorker &worker)
-        : m_worker(worker)
-    {}
-
-    // Inject a read-only sweep observer (W-SWEEP). The worker writes counts as it
-    // schedules/runs sweeps; null in production leaves behavior unchanged.
-    void setSweepObserver(WorkerSweepObserver *observer) { m_worker.m_sweep_observer = observer; }
-
-    // Arm a one-shot fault hook (verification §3, W-THROW). It fires once at the
-    // next fault site reached — the root orchestration body (RunUpdate) or a
-    // failure handler right after it stops the token, before it finishes its
-    // failure bookkeeping — so a test can prove each catch-all contains a throw
-    // and drives a terminal AbortUpdate() without relying on Qt's undefined
-    // slot-throwing behavior. Null in production.
-    void setFaultHook(std::function<void()> hook) { m_worker.m_fault_hook = std::move(hook); }
-
-    // How many per-fetch task handles the worker still holds. This reaches zero
-    // only after every awaited future — including any stopped old-update
-    // straggler — has settled and the deferred sweep has drained, not merely when
-    // an update reaches its terminal state. Ordinary fixtures assert zero at
-    // teardown (verification §2).
-    size_t outstandingFetchTasks() const { return m_worker.m_fetch_tasks.size(); }
-
-    ItemsManagerWorker::ProgressForTest progressCounters() const
-    {
-        return {m_worker.m_stashes_received,
-                m_worker.m_stashes_needed,
-                m_worker.m_characters_received,
-                m_worker.m_characters_needed};
-    }
-
-private:
-    ItemsManagerWorker &m_worker;
 };
